+++ src/main/scala/tel/schich/virtualecu/TestAction.scala
@@ -8,7 +8,7 @@ class TestAction extends Action {
 
     override def execute(t: Long): ActionResult = {
         val f = add(linear(1, 8.minutes), gaussianRandom(-0.001, 0.001)) andThen lowCut(-40) andThen combined(90.0, scale, highCut) andThen shift(40)
-        val continuousResult = f(t)
+        val continuousResult = f(t.toDouble)
         val result = unsignedInteger(1)(continuousResult)
         DataResponse(result)
     }
